Add files via upload
When an MRI or CT is taken at a hospital, it is basically saved as a DCM file and the DCM file can be viewed using a program used in the hospital. This source code converts several dcm files in the folder and subfolders of the original path where the DCM file is stored into PNG files as they are in the destination folder and files.